cleanup: move grpc centric generation code to its own file
This is a standalone change to move gRPC centric code generation
functions to their own file. The changes look large, but all code is
preserved during move; no other changes occur.
