+++ internal/gengapic/BUILD.bazel
@@ -8,6 +8,7 @@ go_library(
         "example.go",
         "generator.go",
         "gengapic.go",
+        "gengrpc.go",
         "genrest.go",
         "helpers.go",
         "imports.go",
