+++ internal/gengapic/client_init.go
@@ -247,180 +247,6 @@ func (g *generator) clientInit(serv *descriptor.ServiceDescriptorProto, servName
 	p("")
 }
 
-func (g *generator) grpcClientInit(serv *descriptor.ServiceDescriptorProto, servName string, imp pbinfo.ImportSpec, hasRPCForLRO bool) {
-	p := g.printf
-
-	// We DON'T want to export the transport layers.
-	lowcaseServName := lowerFirst(servName)
-
-	p("// %sGRPCClient is a client for interacting with %s over gRPC transport.", lowcaseServName, g.apiName)
-	p("//")
-	p("// Methods, except Close, may be called concurrently. However, fields must not be modified concurrently with method calls.")
-	p("type %sGRPCClient struct {", lowcaseServName)
-	p("// Connection pool of gRPC connections to the service.")
-	p("connPool gtransport.ConnPool")
-	p("")
-
-	p("// flag to opt out of default deadlines via %s", disableDeadlinesVar)
-	p("disableDeadlines bool")
-	p("")
-
-	p("// Points back to the CallOptions field of the containing %sClient", servName)
-	p("CallOptions **%sCallOptions", servName)
-	p("")
-
-	p("// The gRPC API client.")
-	p("%s %s.%sClient", grpcClientField(servName), imp.Name, serv.GetName())
-	p("")
-
-	if hasRPCForLRO {
-		p("// LROClient is used internally to handle long-running operations.")
-		p("// It is exposed so that its CallOptions can be modified if required.")
-		p("// Users should not Close this client.")
-		p("LROClient **lroauto.OperationsClient")
-		p("")
-		g.imports[pbinfo.ImportSpec{Name: "lroauto", Path: "cloud.google.com/go/longrunning/autogen"}] = true
-	}
-
-	if g.hasLROMixin() {
-		p("operationsClient longrunningpb.OperationsClient")
-		p("")
-
-		g.imports[pbinfo.ImportSpec{Name: "longrunningpb", Path: "google.golang.org/genproto/googleapis/longrunning"}] = true
-	}
-
-	if g.hasIAMPolicyMixin() {
-
-		p("iamPolicyClient iampb.IAMPolicyClient")
-		p("")
-
-		g.imports[pbinfo.ImportSpec{Name: "iampb", Path: "google.golang.org/genproto/googleapis/iam/v1"}] = true
-	}
-
-	if g.hasLocationMixin() {
-
-		p("locationsClient locationpb.LocationsClient")
-		p("")
-
-		g.imports[pbinfo.ImportSpec{Name: "locationpb", Path: "google.golang.org/genproto/googleapis/cloud/location"}] = true
-	}
-
-	p("// The x-goog-* metadata to be sent with each request.")
-	p("xGoogMetadata metadata.MD")
-
-	p("}")
-	p("")
-
-	g.imports[pbinfo.ImportSpec{Path: "google.golang.org/grpc"}] = true
-	g.imports[pbinfo.ImportSpec{Path: "google.golang.org/grpc/metadata"}] = true
-
-	g.grpcClientUtilities(serv, servName, imp, hasRPCForLRO)
-}
-
-func (g *generator) grpcClientUtilities(serv *descriptor.ServiceDescriptorProto, servName string, imp pbinfo.ImportSpec, hasRPCForLRO bool) {
-	p := g.printf
-
-	clientName := camelToSnake(serv.GetName())
-	clientName = strings.Replace(clientName, "_", " ", -1)
-	lowcaseServName := lowerFirst(servName)
-
-	// Factory function
-	p("// New%sClient creates a new %s client based on gRPC.", servName, clientName)
-	p("//")
-	g.comment(g.comments[serv])
-	p("func New%[1]sClient(ctx context.Context, opts ...option.ClientOption) (*%[1]sClient, error) {", servName)
-	p("  clientOpts := default%[1]sGRPCClientOptions()", servName)
-
-	p("  if new%sClientHook != nil {", servName)
-	p("    hookOpts, err := new%sClientHook(ctx, clientHookParams{})", servName)
-	p("    if err != nil {")
-	p("      return nil, err")
-	p("    }")
-	p("    clientOpts = append(clientOpts, hookOpts...)")
-	p("  }")
-	p("")
-	p("  disableDeadlines, err := checkDisableDeadlines()")
-	p("  if err != nil {")
-	p("    return nil, err")
-	p("  }")
-	p("")
-	p("  connPool, err := gtransport.DialPool(ctx, append(clientOpts, opts...)...)")
-	p("  if err != nil {")
-	p("    return nil, err")
-	p("  }")
-	p("  client := %[1]sClient{CallOptions: default%[1]sCallOptions()}", servName)
-	p("")
-	p("  c := &%sGRPCClient{", lowcaseServName)
-	p("    connPool:    connPool,")
-	p("    disableDeadlines: disableDeadlines,")
-	p("    %s: %s.New%sClient(connPool),", grpcClientField(servName), imp.Name, serv.GetName())
-	p("    CallOptions: &client.CallOptions,")
-	if g.hasLROMixin() {
-		p("    operationsClient: longrunningpb.NewOperationsClient(connPool),")
-	}
-	if g.hasIAMPolicyMixin() {
-		p("    iamPolicyClient: iampb.NewIAMPolicyClient(connPool),")
-	}
-	if g.hasLocationMixin() {
-		p("    locationsClient: locationpb.NewLocationsClient(connPool),")
-	}
-	p("")
-	p("  }")
-	p("  c.setGoogleClientInfo()")
-	p("")
-	p("  client.internal%sClient = c", servName)
-	p("")
-
-	if hasRPCForLRO {
-		p("  client.LROClient, err = lroauto.NewOperationsClient(ctx, gtransport.WithConnPool(connPool))")
-		p("  if err != nil {")
-		p("    // This error \"should not happen\", since we are just reusing old connection pool")
-		p("    // and never actually need to dial.")
-		p("    // If this does happen, we could leak connp. However, we cannot close conn:")
-		p("    // If the user invoked the constructor with option.WithGRPCConn,")
-		p("    // we would close a connection that's still in use.")
-		p("    // TODO: investigate error conditions.")
-		p("    return nil, err")
-		p("  }")
-		p("  c.LROClient = &client.LROClient")
-	}
-
-	p("  return &client, nil")
-	p("}")
-	p("")
-
-	g.imports[pbinfo.ImportSpec{Name: "gtransport", Path: "google.golang.org/api/transport/grpc"}] = true
-	g.imports[pbinfo.ImportSpec{Path: "context"}] = true
-
-	// Connection method
-	p("// Connection returns a connection to the API service.")
-	p("//")
-	p("// Deprecated.")
-	p("func (c *%sGRPCClient) Connection() *grpc.ClientConn {", lowcaseServName)
-	p("  return c.connPool.Conn()")
-	p("}")
-	p("")
-
-	// setGoogleClientInfo method
-	p("// setGoogleClientInfo sets the name and version of the application in")
-	p("// the `x-goog-api-client` header passed on each request. Intended for")
-	p("// use by Google-written clients.")
-	p("func (c *%sGRPCClient) setGoogleClientInfo(keyval ...string) {", lowcaseServName)
-	p(`  kv := append([]string{"gl-go", versionGo()}, keyval...)`)
-	p(`  kv = append(kv, "gapic", versionClient, "gax", gax.Version, "grpc", grpc.Version)`)
-	p(`  c.xGoogMetadata = metadata.Pairs("x-goog-api-client", gax.XGoogHeader(kv...))`)
-	p("}")
-	p("")
-
-	// Close method
-	p("// Close closes the connection to the API service. The user should invoke this when")
-	p("// the client is no longer required.")
-	p("func (c *%sGRPCClient) Close() error {", lowcaseServName)
-	p("  return c.connPool.Close()")
-	p("}")
-	p("")
-}
-
 func (g *generator) makeClients(serv *descriptor.ServiceDescriptorProto, servName string) error {
 	var hasLRO bool
 	for _, m := range serv.GetMethod() {
